Breakout MarkerDrawable
